adk-a2a-deterministic-processing/requirements.txt
google-adk[a2a]>=1.27.0
litellm>=1.80.0
openai>=1.0.0
python-dotenv>=1.0.0
uvicorn>=0.30.0
httpx>=0.27.0
